+++ packages/morpho-sdk/src/actions/requirements/encode/encodeVaultExitBundlesV1Permit.ts
@@ -8,24 +8,59 @@ import {
 import {
   InKindRedeemPermitMismatchError,
   type PermitRequirementSignature,
-} from "../types/index.js";
+} from "../../../types/index.js";
 
-/** Permit tuple consumed by VaultExitBundlesV1. @internal */
+/** Permit tuple consumed by VaultExitBundlesV1. */
 export interface VaultExitBundlesV1Permit {
+  /** Vault-share allowance authorized by the permit. */
   readonly value: bigint;
+  /** Vault permit nonce signed by the owner. */
   readonly nonce: bigint;
+  /** Timestamp after which the permit is invalid. */
   readonly deadline: bigint;
+  /** ECDSA recovery identifier, or zero for the empty-permit sentinel. */
   readonly v: number;
+  /** ECDSA signature `r`, or zero for the empty-permit sentinel. */
   readonly r: `0x${string}`;
+  /** ECDSA signature `s`, or zero for the empty-permit sentinel. */
   readonly s: `0x${string}`;
 }
 
-/** Builds and validates the permit tuple embedded in an in-kind redemption. @internal */
-export const getVaultExitBundlesV1Permit = (params: {
+/** Parameters for {@link encodeVaultExitBundlesV1Permit}. */
+export interface EncodeVaultExitBundlesV1PermitParams {
+  /** Vault share token authorized by the permit. */
   readonly vault: Address;
+  /** Bundle deadline used by the empty-permit sentinel. */
   readonly deadline: bigint;
+  /** Optional signed max-share ERC-2612 requirement. */
   readonly requirementSignature?: PermitRequirementSignature;
-}): VaultExitBundlesV1Permit => {
+}
+
+/**
+ * Encodes an optional vault-share requirement signature as the permit tuple consumed by
+ * VaultExitBundlesV1.
+ *
+ * Without a signature, returns the contract's empty-permit sentinel. With a signature, validates
+ * the ERC-2612 kind, vault asset, and max-share amount before splitting the serialized signature.
+ * Owner, spender, deadline, nonce, and cryptographic validity are verified onchain by the vault.
+ *
+ * @param params - Vault permit encoding parameters.
+ * @returns The VaultExitBundlesV1 permit tuple.
+ * @throws {InKindRedeemPermitMismatchError} when the requirement has the wrong permit kind, asset, amount, or signature encoding.
+ * @example
+ * ```ts
+ * import { encodeVaultExitBundlesV1Permit } from "@morpho-org/morpho-sdk";
+ *
+ * const permit = encodeVaultExitBundlesV1Permit({
+ *   vault,
+ *   deadline,
+ *   requirementSignature,
+ * });
+ * ```
+ */
+export const encodeVaultExitBundlesV1Permit = (
+  params: EncodeVaultExitBundlesV1PermitParams,
+): VaultExitBundlesV1Permit => {
   const { requirementSignature } = params;
   if (requirementSignature == null) {
     return {
