+++ packages/morpho-sdk/src/actions/requirements/encode/index.ts
@@ -2,4 +2,5 @@ export * from "./encodeBlueSignatureAuthorization.js";
 export * from "./encodeErc20Approval.js";
 export * from "./encodeErc20Permit.js";
 export * from "./encodeErc20Permit2Approve.js";
+export * from "./encodeVaultExitBundlesV1Permit.js";
 export * from "./encodeVaultSharesPermit.js";
