+++ packages/morpho-sdk/src/actions/vaultV1/inKindRedeem.ts
@@ -11,7 +11,7 @@ import {
   type Transaction,
   type VaultV1InKindRedeemAction,
 } from "../../types/index.js";
-import { getVaultExitBundlesV1Permit } from "../inKindRedeem.js";
+import { encodeVaultExitBundlesV1Permit } from "../requirements/encode/encodeVaultExitBundlesV1Permit.js";
 
 /** Parameters for {@link vaultV1InKindRedeem}. */
 export interface VaultV1InKindRedeemParams {
@@ -77,7 +77,7 @@ export const vaultV1InKindRedeem = ({
     irm: marketParams.irm,
     lltv: marketParams.lltv,
   }));
-  const sharesPermit = getVaultExitBundlesV1Permit({
+  const sharesPermit = encodeVaultExitBundlesV1Permit({
     vault: vault.address,
     deadline: args.deadline,
     requirementSignature: args.requirementSignature,
